\name{local.copy.package-package}
\alias{local.copy.package-package}
\alias{local.copy.package}
\docType{package}
\title{
  Make a local copy of a package and load it.
}
\description{
  Make a local copy of a package and load
  the copy.  Loading a package this way makes it
  possible to update packages in the common library
  while R sessions have a package attached.  Tries
  to be smart about using existing copies of the package.
}
\seealso{
\code{\link{library.local}()}
}
